+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2025-10-01
+### Added
+- Add weekly cloudLibrary alarm
+
 ## 2025-09-30
 ### Added
 - Add hours and closures v2 alarms